Finish Refactor Transformations
