+++ cvpods/data/transforms/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
-from .transform import *
-from .transform_gen import *
+from .augmentations import *
+from .auto_aug import AutoAugment
 
 __all__ = [k for k in globals().keys() if not k.startswith("_")]
